Extract callbacks to module
