+++ lib/stimulus_reflex.rb
@@ -11,6 +11,7 @@ require "cable_ready"
 require "stimulus_reflex/version"
 require "stimulus_reflex/cable_ready_channels"
 require "stimulus_reflex/configuration"
+require "stimulus_reflex/callbacks"
 require "stimulus_reflex/reflex"
 require "stimulus_reflex/element"
 require "stimulus_reflex/sanity_checker"
